+++ netpyne/batch/utils.py
@@ -7,11 +7,16 @@ from __future__ import unicode_literals
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
+import builtins
 
 from future import standard_library
+
 standard_library.install_aliases()
 
 import numpy as np
+import json
+import pickle
+from subprocess import Popen
 
 # -------------------------------------------------------------------------------
 # function to create a folder if it does not exist
@@ -43,54 +48,46 @@ def createFolder(folder):
 # function to define template for bash submission
 # -------------------------------------------------------------------------------
 
-def bashTemplate(template):
+def jobStringMPIDirect(custom, folder, command):
+    return f"""#!/bin/bash
+{custom}
+cd {folder}
+{command}
     """
-    Function for/to <short description of `netpyne.batch.utils.bashTemplate`>
 
-    Parameters
-    ----------
-    template : <type>
-        <Short description of template>
-        **Default:** *required*
-
-"""
-
-    if template=='mpi_direct':
-        return """#!/bin/bash
-%s
-cd %s
-%s
-        """
-    elif template=='hpc_slurm':
-        return """#!/bin/bash
-#SBATCH --job-name=%s
-#SBATCH -A %s
-#SBATCH -t %s
-#SBATCH --nodes=%d
-#SBATCH --ntasks-per-node=%d
-#SBATCH -o %s.run
-#SBATCH -e %s.err
-#SBATCH --mail-user=%s
+def jobStringHPCSlurm(jobName, allocation, walltime, nodes, coresPerNode, jobPath, email, reservation, custom, folder, command):
+    res = '#SBATCH --res=%s' % (reservation) if reservation else ''
+    return f"""#!/bin/bash
+#SBATCH --job-name={jobName}
+#SBATCH -A {allocation}
+#SBATCH -t {walltime}
+#SBATCH --nodes={nodes}
+#SBATCH --ntasks-per-node={coresPerNode}
+#SBATCH -o {jobPath}.run
+#SBATCH -e {jobPath}.err
+#SBATCH --mail-user={email}
 #SBATCH --mail-type=end
-%s
-%s
+{res}
+{custom}
 source ~/.bashrc
-cd %s
-%s
+cd {folder}
+{command}
 wait
-        """
-    elif template=='hpc_torque':
-        return """#!/bin/bash
-#PBS -N %s
-#PBS -l walltime=%s
-#PBS -q %s
-#PBS -l %s
-#PBS -o %s.run
-#PBS -e %s.err
-%s
+    """
+
+
+def jobStringHPCTorque(jobName, walltime, queueName, nodes, coresPerNode, jobPath, custom, command):
+    return f"""#!/bin/bash
+#PBS -N {jobName}
+#PBS -l walltime={walltime}
+#PBS -q {queueName}
+#PBS -l nodes={nodes}:ppn={coresPerNode}
+#PBS -o {jobPath}.run
+#PBS -e {jobPath}.err
+{custom}
 cd $PBS_O_WORKDIR
 echo $PBS_O_WORKDIR
-%s
+{command}
         """
 
 def cp(obj, verbose=True, die=True):
@@ -114,6 +111,7 @@ def cp(obj, verbose=True, die=True):
         **Options:** ``<option>`` <description of option>
 
 '''
+    from copy import copy
     try:
         output = copy.copy(obj)
     except Exception as E:
@@ -239,3 +237,279 @@ def sigfig(X, sigfigs=5, SI=False, sep=False, keepints=False):
         return tuple(output)
     else:
         return output[0]
+
+
+# func needs to be outside of class
+def runJob(nrnCommand, script, cfgSavePath, netParamsSavePath, simDataPath, rankId):
+    """
+    Function for/to <short description of `netpyne.batch.utils.runJob`>
+
+    Parameters
+    ----------
+    script : <type>
+        <Short description of script>
+        **Default:** *required*
+
+    cfgSavePath : <type>
+        <Short description of cfgSavePath>
+        **Default:** *required*
+
+    netParamsSavePath : <type>
+        <Short description of netParamsSavePath>
+        **Default:** *required*
+
+    simDataPath : <type>
+        <Short description of simDataPath>
+        **Default:** *required*
+
+
+    """
+
+    import os
+    print('\nJob in rank id: ', rankId)
+    command = '%s %s simConfig=%s netParams=%s' % (nrnCommand, script, cfgSavePath, netParamsSavePath)
+    print(command)
+
+    with open(simDataPath+'.run', 'w') as outf, open(simDataPath+'.err', 'w') as errf:
+        pid = Popen(command.split(' '), stdout=outf, stderr=errf, preexec_fn=os.setsid).pid
+
+    with open('./pids.pid', 'a') as file:
+        file.write(str(pid) + ' ')
+
+def evaluator(batch, candidates, args, ngen, pc, **kwargs):
+
+        import os
+        import signal
+        from time import sleep
+
+        total_jobs = 0
+        # options slurm, mpi
+        type = args.get('type', 'mpi_direct')
+
+        # paths to required scripts
+        script = args.get('script', 'init.py')
+        netParamsSavePath =  args.get('netParamsSavePath')
+        genFolderPath = batch.saveFolder + '/gen_' + str(ngen)
+        # mpi command setup
+        nodes = args.get('nodes', 1)
+        paramLabels = args.get('paramLabels', [])
+        coresPerNode = args.get('coresPerNode', 1)
+
+        mpiCommand = args.get('mpiCommand', batch.mpiCommandDefault)
+        nrnCommand = args.get('nrnCommand', 'nrniv')
+
+        numproc = nodes*coresPerNode
+        # slurm setup
+        custom = args.get('custom', '')
+        folder = args.get('folder', '.')
+        email = args.get('email', 'a@b.c')
+        walltime = args.get('walltime', '00:01:00')
+        reservation = args.get('reservation', None)
+        allocation = args.get('allocation', 'csd403') # NSG account
+        # fitness function
+        fitnessFunc = args.get('fitnessFunc')
+        fitnessFuncArgs = args.get('fitnessFuncArgs')
+        if batch.method == 'evol':
+            indexToRerun = args.get('defaultFitness')
+        else:
+            indexToRerun = args.get('maxFitness')
+
+        # read params or set defaults
+        sleepInterval = args.get('sleepInterval', 0.2)
+        # create folder if it does not exist
+        createFolder(genFolderPath)
+
+        # remember pids and jobids in a list
+        pids = []
+        jobids = {}
+
+        def jobPathAndNameForCand(index):
+            if batch.method == 'optuna':
+                jobName = "trial_" + str(ngen)
+            else:
+                jobName = "gen_" + str(ngen) + "_cand_" + str(candidate_index)
+            return jobName, genFolderPath + '/' + jobName
+
+        # create a job for each candidate
+        for candidate_index, candidate in enumerate(candidates):
+            # required for slurm
+            sleep(sleepInterval)
+            # name and path
+            jobName, jobPath = jobPathAndNameForCand(candidate_index)
+            # set initial cfg initCfg
+            if len(batch.initCfg) > 0:
+                for paramLabel, paramVal in batch.initCfg.items():
+                    batch.setCfgNestedParam(paramLabel, paramVal)
+            # modify cfg instance with candidate values
+            print(paramLabels, candidate)
+            for label, value in zip(paramLabels, candidate):
+                print('set %s=%s' % (label, value))
+                batch.setCfgNestedParam(label, value)
+            #self.setCfgNestedParam("filename", jobPath)
+            batch.cfg.simLabel = jobName
+            batch.cfg.saveFolder = genFolderPath
+            # save cfg instance to file
+            cfgSavePath = jobPath + '_cfg.json'
+            batch.cfg.save(cfgSavePath)
+
+            if type=='mpi_bulletin':
+                # ----------------------------------------------------------------------
+                # MPI master-slaves
+                # ----------------------------------------------------------------------
+                pc.submit(runJob, nrnCommand, script, cfgSavePath, netParamsSavePath, jobPath, pc.id())
+                print('-'*80)
+            else:
+                # ----------------------------------------------------------------------
+                # MPI job commnand
+                # ----------------------------------------------------------------------
+
+                if mpiCommand == '':
+                    command = '%s %s simConfig=%s netParams=%s ' % (nrnCommand, script, cfgSavePath, netParamsSavePath)
+                else:
+                    command = '%s -n %d %s -python -mpi %s simConfig=%s netParams=%s ' % (mpiCommand, numproc, nrnCommand, script, cfgSavePath, netParamsSavePath)
+
+                # ----------------------------------------------------------------------
+                # run on local machine with <nodes*coresPerNode> cores
+                # ----------------------------------------------------------------------
+                if type=='mpi_direct':
+                    executer = '/bin/bash'
+                    jobString = jobStringMPIDirect(custom, folder, command)
+                # ----------------------------------------------------------------------
+                # run on HPC through slurm
+                # ----------------------------------------------------------------------
+                elif type=='hpc_slurm':
+                    executer = 'sbatch'
+                    jobString = jobStringHPCSlurm(jobName, allocation, walltime, nodes, coresPerNode, jobPath, email, reservation, custom, folder, command)
+                # ----------------------------------------------------------------------
+                # run on HPC through PBS
+                # ----------------------------------------------------------------------
+                elif type=='hpc_torque':
+                    executer = 'qsub'
+                    queueName = args.get('queueName', 'default')
+                    jobString = jobStringHPCTorque(jobName, walltime, queueName, nodes, coresPerNode, jobPath, custom, command)
+                # ----------------------------------------------------------------------
+                # save job and run
+                # ----------------------------------------------------------------------
+                print('Submitting job ', jobName)
+                print(jobString)
+                print('-'*80)
+                # save file
+                batchfile = '%s.sbatch' % (jobPath)
+                with open(batchfile, 'w') as text_file:
+                    text_file.write("%s" % jobString)
+
+                if type == 'mpi_direct':
+                    with open(jobPath+'.run', 'a+') as outf, open(jobPath+'.err', 'w') as errf:
+                        pids.append(Popen([executer, batchfile], stdout=outf, stderr=errf, preexec_fn=os.setsid).pid)
+                else:
+                    with open(jobPath+'.jobid', 'w') as outf, open(jobPath+'.err', 'w') as errf:
+                        pids.append(Popen([executer, batchfile], stdout=outf, stderr=errf, preexec_fn=os.setsid).pid)
+                #proc = Popen(command.split([executer, batchfile]), stdout=PIPE, stderr=PIPE)
+
+                sleep(0.1)
+                #read = proc.stdout.read()
+                if type == 'mpi_direct':
+                    with open('./pids.pid', 'a') as file:
+                        file.write(str(pids))
+                else:
+                    with open(jobPath+'.jobid', 'r') as outf:
+                        read=outf.readline()
+                    print(read)
+                    if len(read) > 0:
+                        jobid = int(read.split()[-1])
+                        jobids[candidate_index] = jobid
+                    print('jobids', jobids)
+
+            total_jobs += 1
+            sleep(0.1)
+
+        # ----------------------------------------------------------------------
+        # gather data and compute fitness
+        # ----------------------------------------------------------------------
+        if type == 'mpi_bulletin':
+            # wait for pc bulletin board jobs to finish
+            try:
+                while pc.working():
+                    sleep(1)
+                #pc.done()
+            except:
+                pass
+        num_iters = 0
+        jobs_completed = 0
+        fitness = [None for cand in candidates]
+        # print outfilestem
+        if batch.method == 'evol':
+            totalGen = args.get('max_generations')
+        else:
+            totalGen = args.get('maxiters')
+        print(f"Waiting for jobs from generation {ngen}/{totalGen} ...")
+
+        # print "PID's: %r" %(pids)
+        # start fitness calculation
+        while jobs_completed < total_jobs:
+            unfinished = [i for i, x in enumerate(fitness) if x is None ]
+            for candidate_index in unfinished:
+                try: # load simData and evaluate fitness
+                    _, jobPath = jobPathAndNameForCand(candidate_index)
+                    dataPath = jobPath + '_data.json'
+                    simData = None
+                    if os.path.isfile(dataPath):
+                        with open(dataPath) as file:
+                            simData = json.load(file)['simData']
+                    else:
+                        dataPath = jobPath + '_data.pkl'
+                        if os.path.isfile(dataPath):
+                            with open(dataPath, 'rb') as file:
+                                simData = pickle.load(file)['simData']
+                    if simData:
+                        fitness[candidate_index] = fitnessFunc(simData, **fitnessFuncArgs)
+                        jobs_completed += 1
+                        print('  Candidate %d fitness = %.1f' % (candidate_index, fitness[candidate_index]))
+                except Exception as e:
+                    err = "There was an exception evaluating candidate %d:"%(candidate_index)
+                    print(("%s \n %s"%(err,e)))
+            num_iters += 1
+            print('completed: %d' %(jobs_completed))
+            if num_iters >= args.get('maxiter_wait', 5000):
+                print("Max iterations reached, the %d unfinished jobs will be canceled and set to default fitness" % (len(unfinished)))
+                for canditade_index in unfinished:
+                    fitness[canditade_index] = indexToRerun # rerun those that didn't complete;
+                    jobs_completed += 1
+                    try:
+                        if 'scancelUser' in kwargs:
+                            os.system('scancel -u %s'%(kwargs['scancelUser']))
+                        else:
+                            os.system('scancel %d' % (jobids[candidate_index]))  # terminate unfinished job (resubmitted jobs not terminated!)
+                    except:
+                        pass
+            sleep(args.get('time_sleep', 1))
+        # kill all processes
+        if type == 'mpi_bulletin':
+            try:
+                with open("./pids.pid", 'r') as file: # read pids for mpi_bulletin
+                    pids = [int(i) for i in file.read().split(' ')[:-1]]
+                with open("./pids.pid", 'w') as file: # delete content
+                    pass
+                for pid in pids:
+                    try:
+                        os.killpg(os.getpgid(pid), signal.SIGTERM)
+                    except:
+                        pass
+            except:
+                pass
+        elif type == 'mpi_direct':
+            import psutil
+            PROCNAME = "nrniv"
+            for proc in psutil.process_iter():
+                # check whether the process name matches
+                try:
+                    if proc.name() == PROCNAME:
+                        proc.kill()
+                except:
+                    pass
+
+        print("-" * 80)
+        print("  Completed a generation  ")
+        print("-" * 80)
+
+        return fitness